Timer overflow issue

+++ include/devices/devices.h
@@ -31,7 +31,7 @@ device_mode_t device_mode_get();
 
 void device_init();
 
-void device_task(uint32_t timestamp);
+void device_task(uint64_t timestamp);
 
 bool device_start(device_mode_t mode, device_method_t method);
