+++ include/input/macros/macro_shutdown.h
@@ -4,6 +4,6 @@
 #include "input_shared_types.h"
 #include <stdint.h>
 
-void macro_shutdown(uint32_t timestamp, button_data_s *buttons);
+void macro_shutdown(uint64_t timestamp, button_data_s *buttons);
 
 #endif
